+++ lib/components/Layout/Footer.tsx
@@ -1,0 +1,12 @@
+// import { CSSObject } from '@emotion/react';
+// export const Footer = ({ children }) => {
+//   return (
+//     <footer css={baseStyles.footer}>Here's some footer stuff {children}</footer>
+//   );
+// };
+
+// const baseStyles: { [key: string]: CSSObject } = {
+//   footer: {
+//     minHeight: '100px'
+//   }
+// };