+++ lib/components/Modals/Modal.tsx
@@ -1,4 +1,4 @@
-import { FC, ReactNode } from 'react';
+import { FC, ReactNode, useEffect, useState } from 'react';
 import { baseTheme } from '../../styling/baseTheme';
 import { CSSObject } from '@emotion/react';
 import { XmarkCircle } from 'iconoir-react';
@@ -9,18 +9,32 @@ type ModalTypes = {
     modal?: CSSObject;
   };
 
-  close?: () => void;
+  closeModal?: () => void;
 };
 
-export const Modal: FC<ModalTypes> = ({ className = {}, children, close }) => {
+export const Modal: FC<ModalTypes> = ({
+  className = {},
+  children,
+  closeModal
+}) => {
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    setIsVisible(true);
+  }, []);
+
   const closeModalHandler = () => {
-    close?.();
+    closeModal?.();
   };
   return (
-    <div css={[baseTheme.modal, className?.modal]} className="pa-modal">
+    <div
+      css={[baseTheme.modal, className?.modal]}
+      className="pa-modal"
+      style={isVisible ? baseStyles.visible : baseStyles.hidden}
+    >
       <div css={baseTheme.modalContentWrapper}>
         <div css={baseTheme.modalHeading} className="pa-modal-heading">
-          {close && (
+          {closeModal && (
             <XmarkCircle
               style={baseTheme.modalCloseButton}
               onClick={closeModalHandler}
@@ -33,4 +47,14 @@ export const Modal: FC<ModalTypes> = ({ className = {}, children, close }) => {
       </div>
     </div>
   );
+};
+
+const baseStyles = {
+  visible: {
+    opacity: '1',
+    transition: 'opacity .333s'
+  },
+  hidden: {
+    opacity: '0'
+  }
 };